+++ src/components/footer/footer.js
@@ -6,6 +6,7 @@
  * @describe: footer
  */
 import footerTemp from '../../template/footer.html';
+import aplayerTemp from '../../template/aplayer.html';
 import footerImg from './../../images/webp/footer.webp';
 import backgroundImg from './../../images/webp/background.webp';
 import cloudsImg from './../../images/webp/clouds.webp';
@@ -16,10 +17,8 @@ export default function main(_) {
     const footer = $('#footer');
     const footerText = footer.text();
 
-    let footerHtml = footerTemp;
     let config = _.__config.footer;
-
-    footerHtml = _.__tools.tempReplacement(footerHtml, 'footerText', footerText);
+    let footerHtml = _.__tools.tempReplacement(footerTemp, 'footerText', footerText);
 
     /**
      * 设置标语
@@ -60,11 +59,11 @@ export default function main(_) {
     })();
 
     /**
-     * 设置网站统计ID
+     * 设置网站统计ID（已废弃）
      */
-    (() => {
-        if (_.__config.cnzz) footerHtml = _.__tools.tempReplacement(footerHtml, 'cnzzId', _.__config.cnzz);
-    })();
+    // (() => {
+    //     if (_.__config.cnzz) footerHtml = _.__tools.tempReplacement(footerHtml, 'cnzzId', _.__config.cnzz);
+    // })();
 
     /**
      * 添加页脚
@@ -103,32 +102,68 @@ export default function main(_) {
     (() => {
         window.setInterval( () => {
             let runDate = _.__tools.getRunDate(_.__config.info.startDate ? _.__config.info.startDate : '2021-01-01');
-            $('#blogRunTimeSpan').text('This blog has running : '+runDate.daysold+' d '+runDate.hrsold+' h '+runDate.minsold+' m '+runDate.seconds+' s');
+            $('#blogRunTimeSpan').text('博客已运行 : '+runDate.daysold+' 天 '+runDate.hrsold+' 时 '+runDate.minsold+' 分 '+runDate.seconds+' 秒');
         }, 500 );
     })();
 
+    // /**
+    //  * 定时网站统计（已废弃）
+    //  */
+    // (() => {
+    //     if (_.__config.cnzz) {
+    //         _.__timeIds.cnzzTId = window.setInterval( () => {
+    //             let cnzzStat = $('.id_cnzz_stat_icon a');
+    //             if (cnzzStat.length > 0) {
+    //                 let cnzzInfo = [];
+    //                 let cnzzArr  = $(cnzzStat[1]).text().split('|');
+    //                 $.each(cnzzArr, (i) => {
+    //                     let str = $.trim(cnzzArr[i]);
+    //                     if (str !== '') {
+    //                         str = str.replace('今日','Today').replace('昨日','Yesterday').replace('[',':').replace(']','');
+    //                         cnzzInfo.push(str)
+    //                     }
+    //                 });
+    //                 cnzzInfo.push($(cnzzStat[2]).text().replace('当前在线','Online').replace('[',':').replace(']',''));
+    //                 $('#cnzzInfo').text(cnzzInfo.join(' | ')).show();
+    //                 _.__tools.clearIntervalTimeId(_.__timeIds.cnzzTId);
+    //             }
+    //         }, 1000 );
+    //     }
+    // })();
+
     /**
-     * 定时网站统计
+     * 设置音乐播放器
      */
     (() => {
-        if (_.__config.cnzz) {
-            _.__timeIds.cnzzTId = window.setInterval( () => {
-                let cnzzStat = $('.id_cnzz_stat_icon a');
-                if (cnzzStat.length > 0) {
-                    let cnzzInfo = [];
-                    let cnzzArr  = $(cnzzStat[1]).text().split('|');
-                    $.each(cnzzArr, (i) => {
-                        let str = $.trim(cnzzArr[i]);
-                        if (str !== '') {
-                            str = str.replace('今日','Today').replace('昨日','Yesterday').replace('[',':').replace(']','');
-                            cnzzInfo.push(str)
-                        }
-                    });
-                    cnzzInfo.push($(cnzzStat[2]).text().replace('当前在线','Online').replace('[',':').replace(']',''));
-                    $('#cnzzInfo').text(cnzzInfo.join(' | ')).show();
-                    _.__tools.clearIntervalTimeId(_.__timeIds.cnzzTId);
-                }
-            }, 1000 );
+        if (config.aplayer.enable) {
+            Promise.all([_.__tools.dynamicLoadingJs(config.aplayer.cdn.aplayer), _.__tools.dynamicLoadingJs(config.aplayer.cdn.meting)])
+                .then(r => {
+                    _.__tools.dynamicLoadingCss(config.aplayer.cdn.aplayercss);
+
+                    let aplayerHtml = _.__tools.batchTempReplacement(aplayerTemp, [
+                        ['id', config.aplayer.options.id],
+                        ['api', config.aplayer.options.api],
+                        ['server', config.aplayer.options.server],
+                        ['type', config.aplayer.options.type],
+                        ['auto', config.aplayer.options.auto],
+                        ['fixed', config.aplayer.options.fixed],
+                        ['mini', config.aplayer.options.mini],
+                        ['autoplay', config.aplayer.options.autoplay],
+                        ['theme', config.aplayer.options.theme],
+                        ['loop', config.aplayer.options.loop],
+                        ['order', config.aplayer.options.order],
+                        ['preload', config.aplayer.options.preload],
+                        ['volume', config.aplayer.options.volume],
+                        ['mutex', config.aplayer.options.mutex],
+                        ['lrcType', config.aplayer.options.lrcType],
+                        ['listFolded', config.aplayer.options.listFolded],
+                        ['listMaxHeight', config.aplayer.options.listMaxHeight],
+                        ['storageName', config.aplayer.options.storageName],
+                    ]);
+
+                    $('#footer').append(aplayerHtml)
+                })
+                .catch(e => console.error('aplayer|meting', e))
         }
     })();
 }